updated to new fits
updated to produce the average of 6 new fits

In the future we need to make some sort of assumption as to what the council will do if that happens, but for now I assumed that if ABC is so low that the TAC function predicts a negative TAC, that TAC is set to 0.  If, in turn, with a zero TAC the catch function assumes an Catch that exceeds, I'll allow it--we see examples of TAC exceeding ABC on occasion.  It's rare, but it happens.  Chances are we want to go in and program a fishery shutdown if ABC is exceeded, and make some guesses as to what that'd do to the catches of other species but I simply didn't have the time to handle that today.  So for now, we need to put the massive caveat that if the species ABCs go well outside of the observed range the predictions may be very off, as with any prediction based on a regression I suppose.